WIP fix for coreml downloader for whisper-cpp
This won't work in its current form, but I want to push this WIP to
avoid losing it in a system move (in case it helps me or someone else
pick this work back up later).

This work was originally done in late August 2023 atop 2739707f974e
and has somewhat rotted since then as the package has been moved to
by-name, the package has been refactored, and the upstream project
disabled their coreml downloader rather than try to fix it.

I haven't attempted to build this or figure out how my addition to
the download-models.patch should adapt to changes upstream. We also
probably need to upstream a real fix to their downloader if there
isn't an open PR doing it on their repo already.

All I did, really, was copy stuff from a stash to the new files and
write this commit message :)